chore: add context to template functions
