+++ components/actionbutton/stories/actionbutton.stories.js
@@ -1,6 +1,4 @@
 import { default as IconStories } from "@spectrum-css/icon/stories/icon.stories.js";
-import { html } from "lit";
-import { styleMap } from "lit/directives/style-map.js";
 import { Variants } from "./template";
 
 /**
@@ -143,33 +141,7 @@ export default {
 		actions: {
 			handles: ["click .spectrum-ActionButton:not([disabled])"],
 		},
-		html: {
-			root: ".spectrum-ActionButton:first-of-type"
-		}
 	},
-	decorators: [
-		(Story, context) => html`
-			<style>
-				.spectrum-Detail { display: inline-block; }
-				.spectrum-Typography > div {
-					border: 1px solid var(--spectrum-gray-200);
-					border-radius: 4px;
-					padding: 0 14px 14px;
-					/* Why seafoam? Because it separates it from the component styles. */
-					--mod-detail-font-color: var(--spectrum-seafoam-900);
-				}
-			</style>
-			<div style=${styleMap({
-				"display": "flex",
-				"flex-direction": "column",
-				"align-items": "flex-start",
-				"gap": "16px",
-				"--mod-detail-margin-end": "4.8px",
-			})}>
-				${Story(context)}
-			</div>
-		`,
-	],
 };
 
 export const Default = Variants.bind({});
